requirements.txt
# Core dependencies
game-sdk==0.1.3
transformers==4.36.2
torch==2.1.2
pandas==2.1.4
python-dotenv==1.0.0
termcolor==2.4.0

# Twitter API dependencies
httpx==0.26.0

# Async support
asyncio==3.4.3

# Path handling
pathlib==1.0.1

# Data manipulation
numpy==1.24.3

# Frontend
streamlit==1.31.1
plotly==5.18.0

# Optional but recommended for development
black==23.12.1  # Code formatting
pylint==3.0.3   # Code linting
pytest==7.4.3   # Testing

# Optional visualization tools
matplotlib==3.8.2
seaborn==0.13.1
jupyter==1.0.0

# New Web3 dependencies
web3==6.15.1
eth-account==0.10.0
eth-typing==3.5.2
eth-utils==2.3.1
